Do not check azp against OIDC_AUDIENCES
The Authorized Party (azp) field should contain the Client ID of the
party to which the ID Token was issued. [1] It's responsibility of
the *client* to check the azp and since we're not an OIDC client here,
but rather acting as a resource server, we shouldn't care about azp.

This is important for use cases where OpenID provider issues an ID token
for an application (= relaying party, RP) which then authenticates
itself to the resource server (RS) with the same ID token.  In that case
the "aud" field of the ID token should contain RP and RS and "azp" field
should contain RP.

[1] http://openid.net/specs/openid-connect-core-1_0.html#IDToken