Use the built-in Django cache instead of a bespoke one.
The previous implementation had a race condition if fn() took non-negligable
amount of time and the cache-decorated function was called multiple times in
quick succession (e.g. if a client makes multiple requests for resources that
require a userinfo lookup.

Instead of improving the existing implementation, this replaces it with an
implementation backed by Django's cache framework, which should also provide
more flexibility.

If the application doesn't have a cache configured, Django will use an LRU
in-memory cache which respects timeouts, which will provide similar behaviour
to the previous implementation.